Add configurable stopped-typing timer.
For protocols that send updates when the remote user stopped typing, the
stopped-typing timer is not needed. That is, we do not need to manually remove
the typing notice because we will be informed by bitlbee. However,
unfortunately, not all protocols do (or did -- I'm not sure what the current
status of MSN is) this, so we have a stopped-typing timer to automatically
remove the typing notice after some number of seconds.

We can detect some of the protocols that support the stopped-typing
notification, namely oscar-based services and Yahoo. However, we cannot detect
XMPP users, who will send stopped-typing notifications (TYPING 0 CTCP messages).
Thus, we have a conflict between XMPP and MSN: MSN will repeatedly send TYPING 1
messages if the remote user is still typing, leaving us to assume the lack of
TYPING 1 means the remote user stopped typing, but XMPP will send TYPING 1 once
and TYPING 0 when finished (or TYPING 2 when stale). If we make the same
assumption for XMPP as for MSN, then an XMPP remote user's typing notice will
disappear after the (very short) delay of 7 seconds.

This change adds a configurable option, bitlbee_typing_timeout, to allow the
user to adjust the delay after which the notice will be removed. For MSN users,
it should be kept at a low value (e.g., 7 seconds). XMPP users should use
something larger (e.g., 60 seconds or so) so as to try to support both
protocols.

This is an unfortunate necessity, but without a reliable way get notifications
that the remote user has stopped typing for protocols like MSN*, I don't see a
way around this hack. Ideally, this functionality should be abstracted in
bitlbee itself.

* I haven't tested this on MSN recently. I'm merely commenting on why the
  stopped-typing timer was originally created years ago, and making it
  configurable so XMPP users can use this script without their typing notices
  disappearing.