Redrawing fixes.
Force a second redraw via a timer. This fixes a redraw issue when the statusbar
item is high priority and overlayed on top of another statusbar item. With the
second redraw, a TYPING 1 then TYPING 0 (without an event_msg) doesn't redraw
the lower statusbar item, leaving a blank space where text should be.